+ product entity

+++ src/main/java/com/synclab/ecommerce/model/Cart.java
@@ -1,6 +1,8 @@
 package com.synclab.ecommerce.model;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
+
 import javax.persistence.*;
 
 @Entity
@@ -14,19 +16,19 @@ public class Cart implements Serializable {
     @Id
     @Column(name = "cart_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long cartId;
+    private Long cartId;
 
     // @OneToOne(fetch = FetchType.LAZY)
     // @JoinTable(name = "users", joinColumns = @JoinColumn(name = "user_id"),
     // inverseJoinColumns = @JoinColumn(name = "user_id"))
     @Column(name = "user_id")
-    private long userId;
+    private Long userId;
 
     @Column(name = "total_price")
-    private long totalPrice;
+    private BigDecimal totalPrice;
 
     @Column(name = "total_items")
-    private long totalItems;
+    private int totalItems;
 
     // endregion
